requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.26.0
plotly>=5.18.0
requests>=2.31.0
wbgapi>=1.0.12
finnhub-python>=2.4.19
openai>=1.30.0
python-dotenv>=1.0.0
yfinance>=0.2.36
pytrends>=4.9.2
fpdf2>=2.7.0
